pyproject.toml
[tool.poetry]
name = "hollyday_land"
version = "0.1.0"
description = ""
authors = ["Ilya Kramer <ilya@lightricks.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
psycopg2-binary = "^2.9.1"
Pillow = "^8.3.1"
boto3 = "^1.18.16"
requests = "^2.26.0"
Markdown = "^3.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
